refactor(material): move family type guards to dedicated tree-shakable module
The four type guards (isPbrMaterial/isStandardMaterial/isShaderMaterial/
isNodeMaterial) previously lived inside their respective heavy material
modules. The frame-graph geometry renderer dynamic-imports the whole
pbr-material/standard-material namespaces, and Rollup cannot per-export
tree-shake a dynamically imported module, so those scenes retained the
unused guards and dragged in getMaterialFamily (a small bundle-size
regression in the frame-graph scenes).

Move the guards into a dedicated material-guards.ts that uses type-only
imports of the concrete material types, so unused guards cost zero bytes
in every scene. Verified: no built chunk references material-family and
the per-scene bundle manifests are unchanged.

Copilot-Session: 448ce72a-7521-4cb5-a78b-fc89c528d383
