+++ packages/babylon-lite/src/index.ts
@@ -234,10 +234,7 @@ export { createNodeNoColorMaterialView } from "./material/node/no-color-view.js"
 export type { NodeMaterial, NodeInputHandle, ParseNodeMaterialOptions } from "./material/node/node-material.js";
 export { createMaterialView } from "./material/material-view.js";
 export { getMaterialFamily } from "./material/material-family.js";
-export { isPbrMaterial } from "./material/pbr/pbr-material.js";
-export { isStandardMaterial } from "./material/standard/standard-material.js";
-export { isShaderMaterial } from "./material/shader/shader-material.js";
-export { isNodeMaterial } from "./material/node/node-material.js";
+export { isPbrMaterial, isStandardMaterial, isShaderMaterial, isNodeMaterial } from "./material/material-guards.js";
 export { markMaterialUboDirty } from "./material/material-dirty.js";
 export { rebuildMaterial } from "./material/material-rebuild.js";
 export { setSceneImageProcessing } from "./scene/scene-image-processing.js";
